# SeatLookFor-main/Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de SeatLookFor",
        "description": "API para la gestión de eventos y reservas de asientos",
        "contact": {
            "email": "contacto@seatlookfor.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/asientos/{idAsi}/comentarios": {
            "get": {
                "tags": [
                    "Comentarios"
                ],
                "summary": "Listar comentarios de un asiento",
                "operationId": "listarComentariosAsiento",
                "parameters": [
                    {
                        "name": "idAsi",
                        "in": "path",
                        "description": "ID del asiento",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 17
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de comentarios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "idCom": {
                                                "type": "integer"
                                            },
                                            "opinion": {
                                                "type": "string"
                                            },
                                            "valoracion": {
                                                "type": "number"
                                            },
                                            "foto": {
                                                "type": "string",
                                                "nullable": true
                                            },
                                            "idUsu": {
                                                "type": "integer"
                                            },
                                            "idAsi": {
                                                "type": "integer"
                                            },
                                            "idEve": {
                                                "type": "integer",
                                                "nullable": true
                                            },
                                            "usuario": {
                                                "properties": {
                                                    "idUsu": {
                                                        "type": "integer"
                                                    },
                                                    "nombre": {
                                                        "type": "string"
                                                    },
                                                    "apellido": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "El asiento no existe"
                    },
                    "500": {
                        "description": "Error interno"
                    }
                }
            },
            "post": {
                "tags": [
                    "Comentarios"
                ],
                "summary": "Crear un comentario para un asiento",
                "description": "Crea un nuevo comentario y valoración (0‑5) para el asiento indicado. El usuario debe estar autenticado.",
                "operationId": "comentarAsiento",
                "parameters": [
                    {
                        "name": "idAsi",
                        "in": "path",
                        "description": "ID del asiento",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 17
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "opinion",
                                    "valoracion"
                                ],
                                "properties": {
                                    "opinion": {
                                        "type": "string",
                                        "example": "Muy cómodo 😍"
                                    },
                                    "valoracion": {
                                        "type": "number",
                                        "format": "float",
                                        "maximum": 5,
                                        "minimum": 0,
                                        "example": 4.5
                                    },
                                    "foto": {
                                        "description": "Imagen en base64 con cabecera data URI",
                                        "type": "string",
                                        "example": "data:image/jpeg;base64,/9j/4AAQSkZJRgABA...",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Comentario guardado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensaje": {
                                            "type": "string",
                                            "example": "Comentario guardado correctamente"
                                        },
                                        "comentario": {
                                            "properties": {
                                                "idCom": {
                                                    "type": "integer"
                                                },
                                                "opinion": {
                                                    "type": "string"
                                                },
                                                "valoracion": {
                                                    "type": "number"
                                                },
                                                "foto": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "idUsu": {
                                                    "type": "integer"
                                                },
                                                "idAsi": {
                                                    "type": "integer"
                                                },
                                                "idEve": {
                                                    "type": "integer",
                                                    "nullable": true
                                                },
                                                "usuario": {
                                                    "properties": {
                                                        "idUsu": {
                                                            "type": "integer"
                                                        },
                                                        "nombre": {
                                                            "type": "string"
                                                        },
                                                        "apellido": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Usuario no autenticado"
                    },
                    "404": {
                        "description": "El asiento no existe"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error interno"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}